+++ foodfight/src/main/java/com/example/foodfight/upload/UploadRepository.java
@@ -2,6 +2,6 @@ package com.example.foodfight.upload;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UploadRepository extends JpaRepository<Upload, Integer>{
+public interface UploadRepository extends JpaRepository<Upload, Long>{
 
 }
